Extracted RequestParser
